Clean up

+++ services/post.service.js
@@ -1,68 +1,21 @@
-const urlJoin = require('url-join');
 const HumHubImporterMixin = require('./mixins/humhub');
 const SpaceManagerMixin = require('./mixins/space-manager');
 const DiscourseMixin = require('./mixins/discourse');
+const PostFormatterMixin = require('./mixins/post-formatter');
 const CONFIG = require('../config');
-const { convertToIsoString, displayNameToUserName, replaceAsync } = require('../utils');
-
-// Not mapped yet
-// 22 Campus 2019 - Espace dédié au Campus 2019 : préparation, partage de compte-rendus, de photos, etc.
-// 47 Jardin du Léman - Nous soutenons la pratique de « l’Être et du Faire Ensemble » au service d'une transition sociétale qui prend soin de l’humain, des collectifs et de l’environnement. \r\nNous organisons, entre pair, un Espace d’Apprentissage et de Création permettant de cultiver nos liens et nos expériences de Gouvernance Partagée et d’Intelligence Collective.
-// 64 Jardiner le Nous à Nantes - Des espaces pour Jardiner le Nous à Nantes
-const categoriesMapping = {
-  1: 9, // Outils numériques
-  9: 10, // Cercle général
-  12: 24, // La Fabrique des Communs
-  18: 6, // Salle commune
-  19: 25, // Modèle socio-économique
-  34: 26, // Créer et animer un jardin
-  36: 27, // Education et gouvernance partagée
-  37: 28, // Grenoble
-  45: 30, // Aide
-  48: 31, // Jardin Clown et GP
-  53: 11, // Jardin Accompagnement d'organisation
-  55: 12, // Jardin Senso
-  59: 5, // Jardin des Limaces
-  65: 13, // Jardin du Vivre Ensemble
-  76: 14, // Jardin des Dominos
-  99: 32, // Jardin Animons la transition
-  102: 33, // Jardin Recherche & Coopérations
-  108: 29, // Chambéry
-  121: 22 // Campus 2023
-};
-
-const contentContainerMapping = {
-  1: 2, // Outils numériques
-  9: 24, // Cercle général
-  12: 42, // La Fabrique des Communs
-  18: 61, // Salle commune
-  19: 62, // Modèle socio-économique
-  34: 138, // Créer et animer un jardin
-  36: 207, // Education et gouvernance partagée
-  37: 210, // Grenoble
-  45: 318, // Aide
-  48: 402, // Jardin Clown et GP
-  53: 482, // Jardin Accompagnement d'organisation
-  55: 501, // Jardin Senso
-  59: 580, // Jardin des Limaces
-  65: 773, // Jardin du Vivre Ensemble
-  76: 944, // Jardin des Dominos
-  99: 2095, // Jardin Animons la transition
-  102: 2222, // Jardin Recherche & Coopérations
-  108: 2710, // Chambéry
-  121: 3176 // Campus 2023
-};
+const { convertToIsoString, displayNameToUserName } = require('../utils');
+const { categoriesMapping } = require('../mappings');
 
 module.exports = {
   name: 'post',
-  mixins: [SpaceManagerMixin, HumHubImporterMixin, DiscourseMixin],
+  mixins: [SpaceManagerMixin, HumHubImporterMixin, DiscourseMixin, PostFormatterMixin],
   settings: {
     source: {
       humhub: {
         baseUrl: CONFIG.HUMHUB_URL,
         jwtToken: CONFIG.HUMHUB_TOKEN,
-        type: 'post',
-        containerId: 773
+        type: 'post'
+        // containerId: 773
       }
     }
   },
@@ -71,32 +24,31 @@ module.exports = {
       const { cat } = ctx.params;
 
       for (const categoryId of cat ? [cat] : Object.values(categoriesMapping)) {
+        this.logger.info(`Cleaning topics in category ${categoryId}...`);
+
         const category = await this.fetchDiscourse(`/c/${categoryId}/show.json`, {
           headers: {
             'Content-Type': 'application/json'
           }
         });
 
-        console.log('category', category.category);
-
         const numPages = Math.floor(category.category.topic_count / 30);
 
-        console.log('numPages', numPages);
-
         for (let page = numPages; page >= 0; page--) {
-          console.log('page', page);
-
           const topics = await this.fetchDiscourse(`/c/${categoryId}/l/latest.json?page=${page}`, {
             headers: {
               'Content-Type': 'application/json'
             }
           });
 
-          for (const topic of topics.topic_list?.topics) {
-            console.log('Deleting topic', topic.id);
-            await this.fetchDiscourse(`/t/${topic.id}.json`, {
-              method: 'DELETE'
-            });
+          if (topics) {
+            for (const topic of topics.topic_list?.topics) {
+              this.logger.info(`Deleting topic ${topic.id}...`);
+
+              await this.fetchDiscourse(`/t/${topic.id}.json`, {
+                method: 'DELETE'
+              });
+            }
           }
         }
       }
@@ -125,18 +77,11 @@ module.exports = {
       // Only import space "Outils numériques"
       // if (!space || space.id !== 1) return false;
 
-      // console.log('data', data.message)
-
-      // return false;
-
       // Only import mapped categories
-
-      // if (!space || !Object.keys(categoriesMapping).includes(`${space.id}`)) {
-      //   this.logger.warn(`Not migrating post because it is part of ${space?.id}`)
-      //   return;
-      // }
-
-      // console.log('DATA', data);
+      if (!space || !Object.keys(categoriesMapping).includes(`${space.id}`)) {
+        this.logger.warn(`Not migrating post because it is part of ${space?.id}`);
+        return;
+      }
 
       // Split title and message
       const matches = data.message.match(/^## ([^\r]*)\r\n\r\n([\S\s]*)/);
@@ -147,16 +92,15 @@ module.exports = {
 
       const username = displayNameToUserName(data.content.metadata.created_by.display_name);
 
-      // const user = await this.fetchDiscourse(`/u/${username}.json`);
-
-      // if (!user) {
-      //   this.logger.warn(`User ${username} not found, ignoring topic...`);
-      //   return;
-      // }
-
-      const comments = data.content.comments.latest;
+      const user = await this.fetchDiscourse(`/u/${username}.json`);
+      if (!user) {
+        this.logger.warn(`User ${username} not found, ignoring topic...`);
+        return;
+      }
 
       const message = await this.formatMessage(matches[2], data.content.files);
+
+      this.logger.info(`Importing topic "${matches[1]}"...`);
 
       const post = await this.fetchDiscourse(`/posts.json`, {
         method: 'POST',
@@ -173,100 +117,18 @@ module.exports = {
         })
       });
 
-      for (const comment of comments) {
-        await this.postComment(post.topic_id, comment);
-      }
-    },
-    async postComment(topicId, comment, replyToPostId) {
-      if (!topicId) {
-        this.logger.warn(`No topicId provided for comment ${JSON.stringify(comment)}`);
-        return false;
-      }
+      if (post) {
+        const comments = await this.getComments(data.id, 'humhub%5Cmodules%5Cpost%5Cmodels%5CPost');
 
-      const username = displayNameToUserName(comment.createdBy.display_name);
+        if (comments) {
+          this.logger.info(`Importing ${comments.length} comments...`);
 
-      const user = await this.fetchDiscourse(`/u/${username}.json`);
-
-      if (!user) {
-        this.logger.warn(`User ${username} not found, ignoring comment...`);
-        return;
-      }
-
-      const message = await this.formatMessage(comment.message, comment.files);
-
-      const post = await this.fetchDiscourse(`/posts.json`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          'Api-Username': displayNameToUserName(comment.createdBy.display_name)
-        },
-        body: JSON.stringify({
-          topic_id: topicId,
-          raw: message,
-          created_at: convertToIsoString(comment.createdAt),
-          reply_to_post_number: replyToPostId
-        })
-      });
-
-      if (comment.comments) {
-        for (const subComment of comment.comments) {
-          await this.postComment(topicId, subComment, post.id);
-        }
-      }
-    },
-    async formatMessage(message, files) {
-      // Remove extra lines (?)
-      message = message.replace(/(\\\r\n|\r\n|\r|\n)/g, '\r\n');
-
-      // Mentions
-      const mentionsRegex = /\[([^\]]+)\]\(mention:.*\"[^"]+\"\)/gm;
-      message = message.replaceAll(mentionsRegex, (match, p1) => {
-        return '@' + displayNameToUserName(p1);
-      });
-
-      // Images
-      const imagesRegex = /!\[[^\]]*\]\((?<filename>.*?)(?=\"|\))(?<optionalpart>\".*\")?\)/g;
-      message = replaceAsync(message, imagesRegex, async (match, p1) => {
-        const guid = p1.replace('file-guid:', '').trim();
-        const file = files?.find(f => f.guid === guid);
-        if (file) {
-          const message = await this.postImage(file);
-          if (message) return message;
+          for (const comment of comments) {
+            await this.postComment(post.topic_id, comment);
+          }
         } else {
-          this.logger.warn(`Could not post file with GUID ${guid}. No attached file found.`);
+          this.logger.warn(`Could not fetch comments for topic "${matches[1]}"`);
         }
-        return match;
-      });
-
-      return message;
-    },
-    async postImage(file) {
-      const imageUrl = `https://jardiniersdunous.org/api/v1/file/download/${file.id}`;
-
-      this.logger.info(`Downloading image ${imageUrl}...`);
-
-      const response = await fetch(imageUrl, {
-        headers: {
-          Authorization: `Bearer ${this.settings.source.humhub.jwtToken}`
-        }
-      });
-
-      const blob = await response.blob();
-
-      if (response.status === 200) {
-        const formData = new FormData();
-        formData.append('type', 'composer');
-        formData.append('file', blob);
-        formData.append('name', file.file_name);
-        formData.append('synchronous', true);
-
-        const upload = await this.fetchDiscourse('/uploads.json', {
-          method: 'POST',
-          headers: {},
-          body: formData
-        });
-
-        return `![${upload.original_filename}|${upload.width}x${upload.height}](${upload.short_url})`;
       }
     }
   }
